ai_backend/requirements.txt
requests
moralis
flask
flask_cors
langchain_core
langchain_community
uuid
langchain-groq